improve docstrings and maintainability
